@@ -1,5 +1,4 @@
-﻿using Gprc.Server.Extensions;
-using Grpc.Server.Services;
+﻿using Grpc.Server.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -8,58 +7,80 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
 using System;
 using System.IdentityModel.Tokens.Jwt;
+using Grpc.Server.Extensions;
 
-namespace Grpc.Server
+namespace Grpc.Server;
+
+internal class Startup
 {
-    public class Startup
+    private static readonly SymmetricSecurityKey SecurityKey = new(GenerateSecurityKey());
+
+
+    private static byte[] GenerateSecurityKey()
     {
-        private static readonly SymmetricSecurityKey _securityKey = new SymmetricSecurityKey(Guid.NewGuid().ToByteArray());
+        var byteArray = new byte[256];
+        var rnd       = new Random();
+        rnd.NextBytes(byteArray); // Fills the array with random byte values
+        return byteArray;
+    }
 
-        public void ConfigureServices(IServiceCollection services)
+
+    /// <inheritdoc cref="ConfigureServices" />
+    /// <remarks>
+    ///     Must be public.
+    /// </remarks>
+    /// <param name="services"></param>
+    public static void ConfigureServices(IServiceCollection services)
+    {
+        services.AddGrpc();
+
+        services.AddJwtAuthorization();
+        services.AddJwtAuthentication(SecurityKey);
+    }
+
+
+    /// <inheritdoc cref="Configure" />
+    /// <remarks>
+    ///     Must be public.
+    /// </remarks>
+    /// <param name="app"></param>
+    /// <param name="env"></param>
+    public static void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
         {
-            services.AddGrpc();
-
-            services.AddJwtAuthorization();
-            services.AddJwtAuthentication(_securityKey);
+            app.UseDeveloperExceptionPage();
         }
 
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        app.UseRouting();
+
+        app.UseAuthentication();
+        app.UseAuthorization();
+
+        app.UseEndpoints(endpoints =>
         {
-            if (env.IsDevelopment())
+            endpoints.MapGrpcService<UsersService>();
+
+            endpoints.MapGet("/jwt", async context =>
             {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseRouting();
-
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapGrpcService<UsersService>();
-
-                endpoints.MapGet("/jwt", async context =>
-                {
-                    var jwt = GenerateJwt();
-                    await context.Response.WriteAsync(jwt);
-                });
-
-                endpoints.MapGet("/", async context =>
-                {
-                    await context.Response.WriteAsync("Communication with gRPC endpoints must be made through a gRPC client. To learn how to create a client, visit: https://go.microsoft.com/fwlink/?linkid=2086909");
-                });
+                var jwt = GenerateJwt();
+                await context.Response.WriteAsync(jwt);
             });
-        }
 
-        private string GenerateJwt()
-        {
-            var credentials = new SigningCredentials(_securityKey, SecurityAlgorithms.HmacSha256);
-            var header = new JwtHeader(credentials);
-            var payload = new JwtPayload("GrpcServer", "GrpcClient", null, DateTime.Now, DateTime.Now.AddSeconds(60));
-            var token = new JwtSecurityToken(header, payload);
+            endpoints.MapGet("/", async context =>
+            {
+                await context.Response.WriteAsync("Communication with gRPC endpoints must be made through a gRPC client. To learn how to create a client, visit: https://go.microsoft.com/fwlink/?linkid=2086909");
+            });
+        });
+    }
 
-            return new JwtSecurityTokenHandler().WriteToken(token);
-        }
+    private static string GenerateJwt()
+    {
+        var credentials = new SigningCredentials(SecurityKey, SecurityAlgorithms.HmacSha256);
+        var header      = new JwtHeader(credentials);
+        var payload     = new JwtPayload("GrpcServer", "GrpcClient", null, DateTime.Now, DateTime.Now.AddSeconds(60));
+        var token       = new JwtSecurityToken(header, payload);
+
+        return new JwtSecurityTokenHandler().WriteToken(token);
     }
 }
